feature (favorite) : add FavoriteStationService to create
