+++ src/main/java/atdd/favorite/application/dto/CreateFavoriteStationRequestView.java
@@ -1,5 +1,6 @@
 package atdd.favorite.application.dto;
 
+import atdd.favorite.domain.FavoriteStation;
 import lombok.Builder;
 import org.springframework.boot.jackson.JsonComponent;
 
@@ -15,5 +16,17 @@ public class CreateFavoriteStationRequestView {
     public CreateFavoriteStationRequestView(String email, Long stationId) {
         this.stationId = stationId;
         this.email = email;
+    }
+
+    public Long getStationId() {
+        return stationId;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public FavoriteStation toEntity(CreateFavoriteStationRequestView requestView){
+        return FavoriteStation.of(requestView);
     }
 }
